refactor: BoardService에서 createNewBoard 분리
